代码生成完成

+++ src/main/java/com/feilong/im/enums/status/ImUserStatusEnum.java
@@ -11,11 +11,11 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum ImUserStatusEnum {
     // 用户状态（0-离线，1-在线，2-忙碌，3-隐身）
-    OFFLINE(0, "离线"),
-    ONLINE(1, "在线"),
-    BUSY(2, "忙碌"),
-    INVISIBLE(3, "隐身"),
+    OFFLINE((byte)0, "离线"),
+    ONLINE((byte)1, "在线"),
+    BUSY((byte)2, "忙碌"),
+    INVISIBLE((byte)3, "隐身"),
     ;
-    private final Integer id;
+    private final Byte id;
     private final String name;
 }
